Clean up socket thread handling, hopefully fixing network crash
git-svn-id: http://dcplusplus.svn.sourceforge.net/svnroot/dcplusplus/dcplusplus/trunk@1030 f586c4b0-9d0d-0410-bc51-f1fe239209fc

+++ win32/DirectoryListingFrame.cpp
@@ -762,7 +762,7 @@ HTREEITEM DirectoryListingFrame::findFile(const StringSearch& str, HTREEITEM roo
 	changeDir(dir);
 
 	// Check file names in list pane
-	for(int i=0; i<files->size(); i++) {
+	for(size_t i = 0; i < files->size(); i++) {
 		ItemInfo* ii = files->getData(i);
 		if(ii->type == ItemInfo::FILE) {
 			if(str.match(ii->file->getName())) {
